+++ .ci_support/environment-mpich.yml
@@ -3,7 +3,6 @@ channels:
 dependencies:
 - ase =3.23.0
 - atomistics =0.1.32
-- executorlib =0.0.2
 - lammps =2024.06.27=*_mpi_mpich_*
 - mpich =4.2.2
 - numpy =1.26.4
