+++ .ci_support/environment-openmpi.yml
@@ -4,7 +4,6 @@ dependencies:
 - coverage
 - ase =3.23.0
 - atomistics =0.1.32
-- executorlib =0.0.2
 - lammps =2024.06.27=*_mpi_openmpi_*
 - numpy =1.26.4
 - openmpi =4.1.6
