+++ pyproject.toml
@@ -26,7 +26,6 @@ classifiers = [
 dependencies = [
     "ase==3.23.0",
     "atomistics==0.1.32",
-    "executorlib==0.0.2",
     "numpy==1.26.4",
     "pandas==2.2.2",
     "jinja2==3.1.4",
